+++ owncloudTestUtil/src/main/java/com/owncloud/android/testutil/OCSharee.kt
@@ -1,0 +1,32 @@
+/**
+ * ownCloud Android client application
+ *
+ * @author Jorge Aguado Recio
+ *
+ * Copyright (C) 2024 ownCloud GmbH.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.owncloud.android.testutil
+
+import com.owncloud.android.domain.sharing.sharees.model.OCSharee
+import com.owncloud.android.domain.sharing.shares.model.ShareType
+
+val OC_SHAREE = OCSharee(
+    label = "",
+    shareType = ShareType.USER,
+    shareWith = "WhoEver",
+    additionalInfo = "",
+    isExactMatch = true
+)